+++ core/r4300/cop1_w.cpp
@@ -35,7 +35,7 @@ void CVT_S_W()
 {
     if (check_cop1_unusable()) return;
     set_rounding();
-    *reg_cop1_simple[core_cffd] = *((long*)reg_cop1_simple[core_cffs]);
+    *reg_cop1_simple[core_cffd] = *((int32_t*)reg_cop1_simple[core_cffs]);
     PC++;
 }
 
@@ -43,6 +43,6 @@ void CVT_D_W()
 {
     if (check_cop1_unusable()) return;
     set_rounding();
-    *reg_cop1_double[core_cffd] = *((long*)reg_cop1_simple[core_cffs]);
+    *reg_cop1_double[core_cffd] = *((int32_t*)reg_cop1_simple[core_cffs]);
     PC++;
 }
